apps/backend/requirements.txt
# Flask and web server
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# RAG and Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM Integration
google-generativeai==0.3.2

# Natural language processing
nltk==3.8.1
transformers==4.37.2
torch==2.1.2
scikit-learn==1.4.0

# HTTP and API
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
tqdm==4.66.1

# PDF processing
PyPDF2==3.0.1

# XML parsing
lxml==5.1.0 